New changes.

+++ teamapp/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class List(models.Model):
     title = models.CharField('Название', max_length=100)
     description = models.TextField('Описание')
-    rating = models.IntegerField('Рейтинг')
+    rating = models.FloatField('Рейтинг')
 
     def __str__(self):
         return self.title
